数据库tinyint类型转为java Integer类型，与OscarTypeConvert中保持一致

+++ mybatis-plus-generator/src/main/java/com/baomidou/mybatisplus/generator/type/TypeRegistry.java
@@ -44,7 +44,7 @@ public class TypeRegistry {
         typeMap.put(Types.LONGVARBINARY, DbColumnType.BYTE_ARRAY);
         typeMap.put(Types.VARBINARY, DbColumnType.BYTE_ARRAY);
         //byte
-        typeMap.put(Types.TINYINT, DbColumnType.BYTE);
+        typeMap.put(Types.TINYINT, DbColumnType.INTEGER);
         //long
         typeMap.put(Types.BIGINT, DbColumnType.LONG);
         //boolean
